Updated interpreter, fixed parser bugs.
The interpreter is still mosty incomplete.

We have scope now.

To insert stuff into the scope wrap it in a fitting signature object
then insert it into the scope with the signature object as key.

the signature object takes care of proper hashing and equality while
also holding additional information.

For example these two : "void a(int a)" and "int a(int b)" should be the
same signature but have different data.

Fixed some parser bugs that hat to do with interpeting (we often just
assigned the whole token as value)
NOTE: It might be good to assign the token as value (or at least as some
additional parameter) so we keep the Posiiton information.